Remove duplicate donut chart, move used donut chart closer to usage, clean up import paths
