+++ x-pack/plugins/ingest_manager/public/applications/ingest_manager/sections/agent_config/details_page/components/index.ts
@@ -4,5 +4,4 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 export { DatasourcesTable } from './datasources/datasources_table';
-export { DonutChart } from './donut_chart';
 export { EditConfigFlyout } from './edit_config';
