+++ x-pack/plugins/ingest_manager/public/applications/ingest_manager/sections/fleet/components/list_layout.tsx
@@ -19,12 +19,12 @@ import {
 } from '@elastic/eui';
 import { Props as EuiTabProps } from '@elastic/eui/src/components/tabs/tab';
 import { useRouteMatch } from 'react-router-dom';
-import { DonutChart } from '../agent_list_page/components/donut_chart';
-import { useGetAgentStatus } from '../../agent_config/details_page/hooks';
-import { useCapabilities, useLink, useGetAgentConfigs } from '../../../hooks';
 import { WithHeaderLayout } from '../../../layouts';
 import { FLEET_ENROLLMENT_TOKENS_PATH, FLEET_AGENTS_PATH } from '../../../constants';
+import { useCapabilities, useLink, useGetAgentConfigs } from '../../../hooks';
+import { useGetAgentStatus } from '../../agent_config/details_page/hooks';
 import { AgentEnrollmentFlyout } from '../agent_list_page/components';
+import { DonutChart } from './donut_chart';
 
 const REFRESH_INTERVAL_MS = 5000;
